Remove from in your face events and move it to start picker
