+++ client/src/pages/Game/pages/Table/components/InYourFaceEvents/components/InYourFaceEvent.tsx
@@ -16,7 +16,6 @@ import { ProductionDoneEvent } from './ProductionDoneEvent'
 import { StandardProjectBoughtEvent } from './StandardProjectBoughtEvent'
 import { StartingSetupEvent } from './StartingSetupEvent'
 import { TilePlacedEvent } from './TilePlacedEvent'
-import { GameStartedEvent } from './GameStartedEvent'
 
 type Props = {
 	event: GameEvent
@@ -57,8 +56,6 @@ export const InYourFaceEvent = ({ event }: Props) => {
 				return <PlayerMovedDelegateEvent event={event} />
 			case EventType.CommitteePartyActivePolicyActivated:
 				return <CommitteePartyActivePolicyActivatedEvent event={event} />
-			case EventType.Started:
-				return <GameStartedEvent event={event} />
 			default:
 				return null
 		}
