+++ client/src/pages/Game/pages/Table/components/InYourFaceEvents/utils/isInYourFaceEvent.ts
@@ -21,7 +21,6 @@ const PROCESSABLE_EVENTS = [
 	EventType.NewGovernment,
 	EventType.PlayerMovedDelegate,
 	EventType.CommitteePartyActivePolicyActivated,
-	EventType.Started,
 ]
 
 export const isInYourFaceEvent = ({ event }: Params) => {
